pyproject.toml
[tool.poetry]
name = "fastapi-walletauth"
version = "3.2.1"
description = "FastAPI extension for user authentication through signature challenges"
authors = ["mhh <mike.hukiewitz@robotter.ai>"]
readme = "README.md"
packages = [{include = "fastapi_walletauth"}]
keywords = ["FastAPI", "authentication", "signature", "wallet", "ethereum", "solana", "web3", "jwt"]
classifiers = [
    "Environment :: Web Environment",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Security",
    "Typing :: Typed",
]

[tool.poetry.scripts]
walletauth-generate-keys = "fastapi_walletauth.generate_keys:main"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = ">=0.100.0"
pynacl = "^1.5.0"
base58 = ">=2.0.0,<3.0"
eth-account = "^0.9.0"
eth-keys = "^0.4.0"
pyjwt = "^2.8.0"
cryptography = "^41.0.3"
pydantic = "^2.0.0"
pydantic-settings = "^2.7.1"
solders = "^0.26.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.24.1"
pytest-cov = "^4.1.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy-extensions = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
